Entity 컬러명 수정 및 ctl->svc 구문 이동 update구문 추가

+++ src/main/java/Roomchelin/roomchelin/EscaperoomDTO.java
@@ -7,7 +7,7 @@ public record EscaperoomDTO(
         String category,
         String road_address,
         String telephone,
-        double mapx,
-        double mapy
+        Integer mapx,
+        Integer mapy
 ) {
 }
